chatbot_2/chat_api_v1/requirement.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
pydantic==2.8.2
requests==2.32.3
python-dotenv==1.0.1
SQLAlchemy==2.0.35
psycopg2-binary==2.9.9
